ingestion/requirements.txt
docling>=2.0.0
google-generativeai>=0.8.0
qdrant-client>=1.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.0.0
